requirements.txt
PyMuPDF>=1.23.0
scikit-learn>=1.4.0
sklearn-crfsuite>=0.3.0
python-dateutil>=2.9.0
numpy>=1.26.0
scipy>=1.12.0
joblib>=1.3.0
python-docx>=1.1.0
pytest>=8.0.0
python-dotenv>=1.0.0

# LLM / LoRA training (scripts/train_lora.py, scripts/extract_with_groq.py)
transformers>=4.47.0
torch>=2.2.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.11.0
trl>=0.9.0
groq>=0.9.0
json-repair>=0.30.0

# Visualizations & TensorBoard
tensorboard>=2.16.0
matplotlib>=3.8.0
